requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.2
openai==1.10.0
numpy==1.26.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
